Circuit and layout for INA826 variant
Add new strain gauge circuitry based on the INA826. This PCB should be
sent to fab at the same time as the INA126 bias corrected one. Using
this board would be preferred since the 826 is a proper rail to rail
instrumentation amplifier, but the other one is fine if this one has
problems. The INA826 is available in SMD packages only and deadbugs +
breadboard tests aren't really feasible.

Work done so far: schematic and PCB layout, no routing.